Divide resolvers even more
